+++ src/ll/sisyphus.py
@@ -1244,6 +1244,7 @@ class Job:
 		self.maxtime = ns.maxtime
 		self.fork = ns.fork
 		self.noisykills = ns.noisykills
+		self.exit_on_error = ns.exit_on_error
 		self.log2file = ns.log2file
 		self.log2stdout = ns.log2stdout
 		self.log2stderr = ns.log2stderr
@@ -1298,12 +1299,12 @@ class Job:
 
 			if self.repeat:
 				while True:
-					self._handleoneexecution()
-					self._createlogs(False) # Recreate stdin/stdout loggers
+					status = self._handleoneexecution()
 					self._run += 1
 					nextrun = self._calc_nextrun()
 					wait = nextrun - datetime.datetime.now()
 					wait_seconds = wait.total_seconds()
+					self._closelogs(status)
 					if wait_seconds > 0:
 						self.setproctitle("Sleeping")
 						self.log.sisyphus.delay.info(f"Sleeping for {wait} until {nextrun}")
@@ -1311,13 +1312,14 @@ class Job:
 					else:
 						self.log.sisyphus.delay.info(f"Restarting immediately")
 			else:
-				self._handleoneexecution()
+				status = self._handleoneexecution()
+				self._closelogs(status)
 
 			if fcntl is not None:
 				fcntl.flock(f, fcntl.LOCK_UN | fcntl.LOCK_NB)
 
 	def _handleoneexecution(self):
-		# type: () -> None
+		# type: () -> Status
 		self._tasks = [] # type: List[Task]
 		self._loggers = [] # type: List[Logger]
 		self._exceptioncount = 0
@@ -1332,7 +1334,7 @@ class Job:
 		self._getcrontab() # Get crontab
 		self.log = Tag(self._log) # Create tagged logger for files
 		self._delayed_logs = [] # type: Optional[LogList]
-		self._createlogs(True) # Create loggers
+		self._createlogs() # Create loggers
 
 		if self.fork and hasattr(os, "fork"):
 			self._tasks = [self.task("parent", str(os.getpid()))]
@@ -1361,10 +1363,10 @@ class Job:
 				self.process = Process.PARENT
 				self.setproctitle(f"{logmessage} (max time {self.maxtime})")
 				# set a signal to delay CTRL-C handling until the child has cleaned up
-				signal.signal(signal.SIGINT, self._signal_int)
+				signal.signal(signal.SIGINT, self._signal_interupt)
 				# set a signal to wake us up to kill the child process after the maximum runtime
 				if self.maxtime is not None:
-					signal.signal(signal.SIGALRM, self._signal_alarm)
+					signal.signal(signal.SIGALRM, self._signal_timeout)
 					signal.alarm(int(self.maxtime.total_seconds()))
 				try:
 					(pid, status) = os.waitpid(pid, 0) # Wait for the child process to terminate
@@ -1399,7 +1401,7 @@ class Job:
 							raise exc
 					elif status is Status.SUCCESSFUL:
 						self._finished_successful(None)
-					return # finish normally (or continue, if we're in repeat mode)
+					return status # finish normally (or continue, if we're in repeat mode)
 			# Here we are in the child process
 			self.process = Process.CHILD
 			self.setproctitle()
@@ -1409,7 +1411,7 @@ class Job:
 		else: # We didn't fork
 			# set a signal to kill ourselves after the maximum runtime
 			if self.maxtime is not None and hasattr(signal, "SIGALRM"):
-				signal.signal(signal.SIGALRM, self._signal_alarm)
+				signal.signal(signal.SIGALRM, self._signal_timeout)
 				signal.alarm(int(self.maxtime.total_seconds()))
 
 		self.setproctitle("Setting up")
@@ -1443,6 +1445,7 @@ class Job:
 		self.notifyfinish(result)
 		if self.fork:
 			os._exit(status)
+		return status
 
 	def _kill_children(self):
 		# type: () -> Set[int]
@@ -1473,6 +1476,15 @@ class Job:
 				# Ignore whether any processes remain in the ``alive`` list
 			return pids
 
+	def _termination_message(self, exc, pids):
+		if not pids:
+			return f"Terminated: {exc}"
+		elif len(pids) == 1:
+			return f"Terminated child {misc.first(pids)}: {exc}"
+		else:
+			pidstr = ", ".join(str(pid) for pid in pids)
+			return f"Ternminated children {pidstr}: {exc}"
+
 	def _finished_uneventful(self):
 		# type: () -> Status
 		self.endtime = datetime.datetime.now()
@@ -1484,7 +1496,6 @@ class Job:
 				self.log.sisyphus.result.errors(None)
 			else:
 				self.log.sisyphus.result.delay.ok(None)
-		self._closelogs(Status.UNEVENTFUL)
 		return Status.UNEVENTFUL
 
 	def _finished_successful(self, result):
@@ -1501,7 +1512,6 @@ class Job:
 				if self._delayed_logs is not None:
 					self._delayed_logs = []
 				self.log.sisyphus.result.ok(result)
-		self._closelogs(Status.SUCCESSFUL)
 		return Status.SUCCESSFUL
 
 	def _finished_exception(self, exc):
@@ -1514,7 +1524,6 @@ class Job:
 			# log the error to the logfile, as we assume that :meth:`execute` didn't do it
 			self.log.sisyphus.email.mattermost(exc)
 			self.log.sisyphus.result.fail(f"failed with {strexc}")
-		self._closelogs(Status.FAILED)
 		return Status.FAILED
 
 	def _finished_break(self, exc):
@@ -1526,7 +1535,6 @@ class Job:
 			# Don't log to email or mattermost
 			self.log.sisyphus(exc)
 			self.log.sisyphus.result.fail(f"failed with {misc.format_exception(exc)}")
-		self._closelogs(Status.INTERRUPTED)
 		return Status.INTERRUPTED
 
 	def _finished_timeout(self, exc):
@@ -1538,32 +1546,22 @@ class Job:
 
 		if self.process is Process.PARENT:
 			pids = self._kill_children()
-
-			if len(pids) == 1:
-				pidstr = f"child {misc.first(pids)}"
-			else:
-				pidstr = ", ".join(str(pid) for pid in pids)
-				pidstr = f"children {pidstr}"
-
-			msg = f"Terminated {pidstr} after {self.maxtime}"
 		elif self.process is Process.SOLO:
-			msg = f"Terminated after {self.maxtime}"
-		else:
-			msg = None
-		if msg is not None:
+			pids = set()
+
+		if self.process is not Process.CHILD:
 			if self.noisykills:
 				self.log.email.mattermost(exc)
 			else:
 				self.log(exc)
-			self.log.sisyphus.result.kill(msg)
-		self._closelogs(Status.TIMEOUT)
+			self.log.sisyphus.result.kill(self._termination_message(exc, pids))
 		return Status.TIMEOUT
 
-	def _signal_alarm(self, signum, frame):
+	def _signal_timeout(self, signum, frame):
 		# type: (int, Optional[types.FrameType]) -> NoReturn
 		raise misc.Timeout(self.maxtime)
 
-	def _signal_int(self, signum, frame):
+	def _signal_interupt(self, signum, frame):
 		# type: (int, Optional[types.FrameType]) -> NoReturn
 		signal.alarm(0) # Cancel maximum runtime alarm
 		# Give the child process time to log the stacktrace
@@ -1773,21 +1771,18 @@ class Job:
 			cutoff = datetime.datetime.now() - cutoff
 		return cutoff
 
-	def _createlogs(self, full):
-		# type: (bool) -> None
+	def _createlogs(self):
+		# type: () -> None
 		"""
 		Create the logfile and the link to the logfile (if configured).
-
-		If ``full`` is false, only the loggers for `stdout` and `stderr` will
-		be generated (if configured).
 		"""
 		self._loggers = []
 		skipfilenames = [] # type: List[pathlib.Path]
-		if full and self.toemail and self.fromemail and self.smtphost:
+		if self.toemail and self.fromemail and self.smtphost:
 			# Use the email logger as the first logger, so that when sending the email (in :meth:`EmailLogger.close`) fails,
 			# it will still be logged to the log file/stdout/stderr
 			self._loggers.append(EmailLogger(self))
-		if full and self.log2file:
+		if self.log2file:
 			logfilename = self.logfilename()
 			if logfilename is not None:
 				# Create the logger for the log file
@@ -1818,14 +1813,27 @@ class Job:
 
 	def _closelogs(self, status):
 		# type: (Status) -> None
-		while self._loggers:
+
+		# Note that in forking mode the child process inherits the delayed log
+		# messages of the parent process. If both processes would log a
+		# non-delayed message, the inherited messages would be output twice.
+		# To avoid this problem, we clear the delayed log queue in the parent
+		# before continuing.
+		if self._delayed_logs:
+			self._delayed_logs = []
+		index = 0
+		while index < len(self._loggers):
 			# Don't remove the logger from the list immediately
 			# In this way, log messages that the logger outputs during closing will
 			# be logged by the logger itself (i.e. logfile cleanup will be logged
 			# in the logfile)
-			logger = self._loggers[0]
-			logger.close(status)
-			del self._loggers[0]
+			logger = self._loggers[index]
+			if logger.close(status):
+				# Logger has closed, so remove it
+				del self._loggers[index]
+			else:
+				# Logger didn't close, keep it and go to the next one
+				index += 1
 
 	def _write_healthfile(self, error):
 		# type: (OptStr) -> None
@@ -1999,11 +2007,14 @@ class Logger:
 		"""
 
 	def close(self, status):
-		# type: (Status) -> None
+		# type: (Status) -> bool
 		"""
 		Called by the :class:`Job` when job execution has finished.
 
 		``status`` (an :class:`Status`) is the result status of the job run.
+
+		Return whether the logfile has been closed. (All normal loggers
+		will close except ``stdout`` and ``stderr`` loggers).
 		"""
 
 
@@ -2036,6 +2047,10 @@ class StreamLogger(Logger):
 			self.stream.write("\n")
 			self.lineno += 1
 		self.stream.flush()
+
+	def close(self, status):
+		# type: (Status) -> bool
+		return False
 
 
 class FileLogger(StreamLogger):
@@ -2098,16 +2113,17 @@ class FileLogger(StreamLogger):
 									self.job.log.sisyphus.delay.info(f"Compressing logfiles older than {compressfilelogs} via {self.job.compressmode}")
 									compressedany = True
 								self.compress(filename)
-			if status is Status.UNEVENTFUL:
-				self.job.log.sisyphus.delay.info("Going to delete current logfile")
-			if removedany or compressedany or status in (Status.UNEVENTFUL, Status.SUCCESSFUL):
-				self.job.log.sisyphus.delay.info("Logfiles cleaned up")
+			if removedany or compressedany:
+				self.job.log.sisyphus.delay.info("Old logfiles cleaned up")
+		if self.job.process is not Process.CHILD and status is Status.UNEVENTFUL:
+			self.job.log.sisyphus.delay.info("Going to delete current logfile")
 		# Close the stream now, so that we're able to delete it (even on Windows)
 		self.stream.close()
 		if self.job.process is not Process.CHILD:
 			if status is Status.UNEVENTFUL:
 				# Remove current log file in case of a uneventful run
 				self.filename.unlink()
+		return True
 
 	def remove(self, filename):
 		# type: (pathlib.Path) -> None
@@ -2173,6 +2189,10 @@ class LinkLogger(Logger):
 			linkname.unlink()
 			linkname.symlink_to(filename)
 
+	def close(self, status):
+		# type: (Status) -> bool
+		return True
+
 
 class CurrentLinkLogger(LinkLogger):
 	"""
@@ -2199,9 +2219,10 @@ class LastStatusLinkLogger(LinkLogger):
 		return f"<{self.__class__.__module__}.{self.__class__.__qualname__} linkname={str(self.linkname)!r} status={self.status.name} at {id(self):#x}>"
 
 	def close(self, status):
-		# type: (Status) -> None
+		# type: (Status) -> bool
 		if self.job.process is not Process.CHILD and status is self.status:
 			self._makelink()
+		return True
 
 
 class EmailLogger(Logger):
@@ -2259,7 +2280,7 @@ class EmailLogger(Logger):
 			pass
 
 	def close(self, status):
-		# type: (Status) -> None
+		# type: (Status) -> bool
 		if self.file is not None:
 			self.file.close()
 		else:
@@ -2365,6 +2386,7 @@ class EmailLogger(Logger):
 					self.job.emailfilename(p).unlink()
 				except FileNotFoundError:
 					pass
+		return True
 
 
 class MattermostLogger(Logger):
@@ -2421,6 +2443,10 @@ class MattermostLogger(Logger):
 					"message": message[:15000],
 				}
 			)
+
+	def close(self, status):
+		# type: (Status) -> bool
+		return True
 
 
 ###